projects/210110B-GNN_Citation-Networks/requirements.txt
numpy>=1.23.0,<2.0
cython>=0.29.36,<3.0
paddlepaddle==2.6.1
pgl>=2.0.0
tqdm>=4.60.0
PyYAML>=5.4.0
easydict>=1.9.0
tensorboardX>=2.2.0
ogb>=1.3.0
tqdm